# Improve performance of tags functions